Fixes derp that broke Overlays on last PR.
I tested Overlays, Dye Colors, Illumination and Hammer interaction. It
seems to work the same as 3.3.7 now.

I was not able to get Covers to work like it says in the Tutorial, but
then again it did not work in 3.3.7 either, so I think that might be
another problem or me not understanding something.

Try giving this anothoer spin.